+++ spec/helpers/form_helper_spec.rb
@@ -9,7 +9,8 @@ RSpec.describe FormHelper, type: :helper do
   end
 
   describe "#list_of_delivery_partners" do
-    it "asks for a list of organisations that are not `service_owner`" do
+    it "asks for a list of organisations that are delivery partners" do
+      _beis = create(:beis_organisation)
       delivery_partner_1 = create(:delivery_partner_organisation, name: "aaaaa")
       delivery_partner_2 = create(:delivery_partner_organisation, name: "zzzzz")
 
@@ -18,6 +19,23 @@ RSpec.describe FormHelper, type: :helper do
 
       expect(helper.list_of_delivery_partners).to match_array([
         delivery_partner_1,
+        delivery_partner_2,
+      ])
+    end
+  end
+
+  describe "#list_of_reporting_organisations" do
+    it "asks for a list of organisations that are delivery partners or the `service_owner`" do
+      beis = create(:beis_organisation)
+      delivery_partner_1 = create(:delivery_partner_organisation, name: "aaaaa")
+      delivery_partner_2 = create(:delivery_partner_organisation, name: "zzzzz")
+
+      _matched_effort_provider = create(:matched_effort_provider)
+      _external_income_provider = create(:external_income_provider)
+
+      expect(helper.list_of_reporting_organisations).to match_array([
+        delivery_partner_1,
+        beis,
         delivery_partner_2,
       ])
     end